add reinstall options for _mini*-start.sh

+++ openshift/_minishift-start.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
-rm -rf ~/.kube ~/.minishift
 if [[ "$OPENSHIFT_VERSION" == "" ]]; then
 	echo "source _minishift-boot.env first before running this script."
-	echo "Ignore errors re: not finding Openshift binary (cuz Openshift is not running yet, right?)"
 	exit -1
+fi
+if [[ $1 == reinstall ]]; then
+  minishift -f delete
+  rm -rf ~/.kube ~/.minishift
 fi
 minishift start --memory $MINISHIFT_VM_MEMORY --vm-driver virtualbox --show-libmachine-logs --openshift-version $OPENSHIFT_VERSION
 echo ""
